requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.1
astor==0.8.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
bidict==0.22.1
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.4
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
docopt==0.6.2
efficientnet-pytorch==0.6.3
entrypoints==0.4
eventlet==0.33.3
executing==1.2.0
filelock==3.12.2
Flask==2.2.3
Flask-SocketIO==5.3.3
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
greenlet==2.0.2
huggingface-hub==0.15.1
idna==3.4
imageio==2.31.1
install==1.3.5
ipykernel==6.22.0
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.0
kiwisolver==1.4.4
lazy_loader==0.2
lit==16.0.5.post0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
munch==3.0.0
nest-asyncio==1.5.6
networkx==3.1
num2words==0.5.12
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==21.3
paddle-bfloat==0.1.7
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.5.4
pretrainedmodels==0.7.4
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pydub==0.25.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-engineio==4.4.0
python-socketio==5.8.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qudida==0.0.4
requests==2.31.0
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@f58f1d1cd1b790b0d5214ae6e80dfbb1d4f19d39
segmentation-models-pytorch==0.2.0
six==1.16.0
stack-data==0.6.2
sympy==1.12
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.4.12
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.65.0
traitlets==5.6.0
triton==2.0.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.5
Werkzeug==2.2.3
yarl==1.8.2
